+++ mongo_rep_admin.py
@@ -486,32 +486,6 @@ def chk_rep_lag(repset, args_array, **kwargs):
                     class_cfg=mongo_cfg, **kwargs)
 
 
-def setup_mail(to_line, subj=None, frm_line=None, **kwargs):
-
-    """Function:  setup_mail
-
-    Description:  Initialize a mail instance.  Provide 'from line' if one is
-        not passed.
-
-    Arguments:
-        (input) to_line -> Mail to line.
-        (input) subj -> Mail subject line.
-        (input) frm_line -> Mail from line.
-        (output) Mail instance.
-
-    """
-
-    to_line = list(to_line)
-
-    if isinstance(subj, list):
-        subj = list(subj)
-
-    if not frm_line:
-        frm_line = getpass.getuser() + "@" + socket.gethostname()
-
-    return gen_class.Mail(to_line, subj, frm_line)
-
-
 def run_program(args_array, func_dict, **kwargs):
 
     """Function:  run_program